pure refactor: put everything latest-at in its own dedicated corner
